Update user status
