+++ app/components/channel_drawer/channels_list/channels_list.js
@@ -13,11 +13,9 @@ import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
 import SearchBar from 'app/components/search_bar';
 import {changeOpacity, makeStyleSheetFromTheme} from 'app/utils/theme';
 
-import Profile from './profile';
 import FilteredList from './filtered_list';
 import List from './list';
 import SwitchTeamsButton from './switch_teams_button';
-import LocalConfig from 'assets/config';
 
 class ChannelsList extends React.PureComponent {
     static propTypes = {
@@ -143,11 +141,6 @@ class ChannelsList extends React.PureComponent {
                                 onShowTeams={onShowTeams}
                             />
                         </View>
-                        {LocalConfig.HideSetStatusExperimental && (
-                            <Profile
-                                navigator={navigator}
-                            />
-                        )}
                         {title}
                     </View>
                 </View>
