+++ assets/base/config.json
@@ -17,7 +17,6 @@
     "HideGitLabLoginExperimental": false,
     "HideLDAPLoginExperimental": false,
     "HideSAMLLoginExperimental": false,
-    "HideSetStatusExperimental": true,
 
     "SentryEnabled": false,
     "SentryDsnIos": "",
